crawler/requirements.txt
requests==2.31.0
pandas==2.1.4
playwright==1.40.0
google-generativeai==0.3.2
python-dotenv==1.0.0
mysql-connector-python==8.2.0
asyncio==3.4.3
